refactored benchmark class inheritance to implement an execution plan that can run multiple frameworks, benchmarks and so on
