+++ amlb/__init__.py
@@ -2,11 +2,11 @@
 amlb entrypoint package.
 """
 
+from .benchmark import SetupMode
 from .logger import app_logger as log
 from .errors import AutoMLError
 from .resources import Resources
-from .benchmark import Benchmark, SetupMode
-from .runners import AWSBenchmark, DockerBenchmark, SingularityBenchmark
+from .runners import LocalBenchmark, AWSBenchmark, DockerBenchmark, SingularityBenchmark
 from .results import TaskResult
 from .__version__ import __version__
 
@@ -14,10 +14,10 @@ __all__ = [
     "log",
     "AutoMLError",
     "Resources",
-    "Benchmark",
+    "LocalBenchmark",
+    "AWSBenchmark",
     "DockerBenchmark",
     "SingularityBenchmark",
-    "AWSBenchmark",
     "SetupMode",
     "TaskResult",
     "__version__"
