+++ amlb/runners/__init__.py
@@ -2,12 +2,14 @@
 benchmark runners
 """
 
+from .local import LocalBenchmark
 from .aws import AWSBenchmark
 from .docker import DockerBenchmark
 from .singularity import SingularityBenchmark
 
 __all__ = [
+    "LocalBenchmark",
+    "AWSBenchmark",
     "DockerBenchmark",
     "SingularityBenchmark",
-    "AWSBenchmark",
 ]
